Token Ready.

+++ SOA/views.py
@@ -1,5 +1,5 @@
 from background_task.models import Task
-from django.contrib.auth import login
+from django.contrib.auth import login, logout
 from django.db import transaction
 from django.http import FileResponse
 from rest_framework import status
@@ -14,11 +14,6 @@ from .tasks import *
 
 # noinspection PyMethodMayBeStatic
 class Signup(APIView):
-    # fixme
-    # for testing frontend purposes Only
-    authentication_classes = []
-    permission_classes = (AllowAny,)
-
     @transaction.atomic
     def post(self, request):
         try:
@@ -31,14 +26,14 @@ class Signup(APIView):
                 user_serializer = UserSerializer(data=data)
                 user_serializer.is_valid(raise_exception=True)
                 user = user_serializer.save()
-                # fixme
-                # for testing frontend purposes Only
-                # log = Log(user=request.user, operation='+', operand='User')
-                # log.add_or_remove_fields(user)
-                # log.save()
-                return Response(status=status.HTTP_200_OK)
+                log = Log(user=request.user, operation='+', operand='User')
+                log.add_or_remove_fields(user)
+                log.save()
+                return Response({'pk': user.id}, status=status.HTTP_200_OK)
         except serializers.ValidationError as e:
-            return Response(e.detail, status=status.HTTP_400_BAD_REQUEST)
+            return Response(e.detail, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
+        except KeyError:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
 
 
 # noinspection PyMethodMayBeStatic
@@ -53,21 +48,32 @@ class Login(APIView):
             user = User.objects.get(username=username)
             if user.check_password(password):
                 login(request, user)
+                if not request.data['keep']:
+                    request.session.set_expiry(0)
                 serializer = UserLoginSerializer(instance=user)
                 return Response(serializer.data)
             else:
                 return Response(status=status.HTTP_401_UNAUTHORIZED)
-        except User.DoesNotExist:
+        except User.DoesNotExist or KeyError:
             return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 
 # noinspection PyMethodMayBeStatic
-class AddArash(APIView):
-    # fixme
-    # for testing frontend purposes Only
-    authentication_classes = []
-    permission_classes = (AllowAny,)
+class GetProfile(APIView):
+    def get(self, request):
+        serializer = UserLoginSerializer(instance=request.user)
+        return Response(serializer.data)
 
+
+# noinspection PyMethodMayBeStatic
+class LogOut(APIView):
+    def get(self, request):
+        logout(request)
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+# noinspection PyMethodMayBeStatic
+class AddArash(APIView):
     @transaction.atomic
     def post(self, request):
         try:
@@ -75,36 +81,27 @@ class AddArash(APIView):
                 serializer = ArashSerializer(data=request.data)
                 serializer.is_valid(raise_exception=True)
                 arash = serializer.save()
-                # fixme
-                # for testing frontend purposes Only
-                # log = Log(user=request.user, operation='+', operand='Arash')
-                # log.add_or_remove_fields(arash)
-                # log.save()
+                log = Log(user=request.user, operation='+', operand='Arash')
+                log.add_or_remove_fields(arash)
+                log.save()
                 check_arash_active(arash.pk, repeat=Task.DAILY, repeat_until=None)
-                return Response(status=status.HTTP_200_OK)
+                return Response(status=status.HTTP_201_CREATED)
         except serializers.ValidationError as e:
-            return Response(e.detail, status=status.HTTP_400_BAD_REQUEST)
+            return Response(e.detail, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
 
 # noinspection PyMethodMayBeStatic, PyUnusedLocal, DuplicatedCode
 class ArashOperations(APIView):
-    # fixme
-    # for testing frontend purposes Only
-    authentication_classes = []
-    permission_classes = (AllowAny,)
-
     @transaction.atomic
     def delete(self, request, pk):
         try:
             with transaction.atomic():
                 arash = Arash.objects.get(pk=pk)
-                # fixme
-                # for testing frontend purposes Only
-                # log = Log(operation='-', operand='Arash', user=request.user)
-                # log.add_or_remove_fields(arash)
-                # log.save()
+                log = Log(operation='-', operand='Arash', user=request.user)
+                log.add_or_remove_fields(arash)
+                log.save()
                 arash.delete()
-                return Response(status=status.HTTP_200_OK)
+                return Response(status=status.HTTP_204_NO_CONTENT)
         except Arash.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
@@ -121,46 +118,32 @@ class ArashOperations(APIView):
         try:
             with transaction.atomic():
                 arash = Arash.objects.get(pk=pk)
+                old_arash = Arash.objects.get(pk=pk)
                 serializer = ArashSerializer(arash, request.data, partial=True)
                 serializer.is_valid(raise_exception=True)
                 new_arash = serializer.save()
-                # fixme
-                # for testing frontend purposes Only
-                # log = Log(operation='*', operand='Arash', user=request.user)
-                # log.edit_fields(arash, new_arash)
-                # log.save()
-                return Response(serializer.data, status=status.HTTP_200_OK)
+                log = Log(operation='*', operand='Arash', user=request.user)
+                log.edit_fields(old_arash, new_arash)
+                log.save()
+                return Response(status=status.HTTP_204_NO_CONTENT)
         except serializers.ValidationError as e:
-            return Response(e.detail, status=status.HTTP_400_BAD_REQUEST)
+            return Response(e.detail, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
         except Arash.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
 
 class GetCompanies(ListAPIView):
-    # fixme
-    # for testing frontend purposes Only
-    authentication_classes = []
-    permission_classes = (AllowAny,)
     queryset = Company.objects.all()
     serializer_class = GetCompanySerializer
 
 
 class GetArashes(ListAPIView):
-    # fixme
-    # for testing frontend purposes Only
-    authentication_classes = []
-    permission_classes = (AllowAny,)
     serializer_class = ArashSerializer
     queryset = Arash.objects.all()
 
 
 # noinspection PyMethodMayBeStatic, PyUnusedLocal, DuplicatedCode
 class CompanyOperations(APIView):
-    # fixme
-    # for testing frontend purposes Only
-    authentication_classes = []
-    permission_classes = (AllowAny,)
-
     def get(self, request, pk):
         try:
             company = Company.objects.get(pk=pk)
@@ -174,13 +157,11 @@ class CompanyOperations(APIView):
         try:
             with transaction.atomic():
                 company = Company.objects.get(pk=pk)
-                # fixme
-                # for testing frontend purposes Only
-                # log = Log(operation='-', operand='Company', user=request.user)
-                # log.add_or_remove_fields(company)
-                # log.save()
+                log = Log(operation='-', operand='Company', user=request.user)
+                log.add_or_remove_fields(company)
+                log.save()
                 company.save()
-                return Response(status=status.HTTP_200_OK)
+                return Response(status=status.HTTP_204_NO_CONTENT)
         except Company.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
@@ -189,6 +170,7 @@ class CompanyOperations(APIView):
         try:
             with transaction.atomic():
                 company = Company.objects.get(pk=pk)
+                old_company = Company.objects.get(pk=pk)
                 address = company.address
                 address_serializer = AddressSerializer(address, request.data['address'])
                 address_serializer.is_valid(raise_exception=True)
@@ -197,14 +179,12 @@ class CompanyOperations(APIView):
                 serializer = CompanySerializer(company, request.data, partial=True)
                 serializer.is_valid(raise_exception=True)
                 new_company = serializer.save()
-                # fixme
-                # for testing frontend purposes Only
-                # log = Log(operation='*', operand='Company', user=request.user)
-                # log.edit_fields(company, new_company)
-                # log.save()
-                return Response(serializer.data, status=status.HTTP_200_OK)
+                log = Log(operation='*', operand='Company', user=request.user)
+                log.edit_fields(old_company, new_company)
+                log.save()
+                return Response(status=status.HTTP_204_NO_CONTENT)
         except serializers.ValidationError as e:
-            return Response(e.detail, status=status.HTTP_400_BAD_REQUEST)
+            return Response(e.detail, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
         except Company.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
@@ -252,12 +232,6 @@ class RequestOperations(APIView):
 
 # noinspection PyMethodMayBeStatic
 class AddCompany(APIView):
-    # permission_classes = [IsAuthenticated, ]
-    # fixme
-    # for testing frontend purposes Only
-    authentication_classes = []
-    permission_classes = (AllowAny,)
-
     @transaction.atomic
     def post(self, request):
         try:
@@ -269,78 +243,69 @@ class AddCompany(APIView):
                 serializer = CompanySerializer(data=request.data)
                 serializer.is_valid(raise_exception=True)
                 company = serializer.save()
-                # fixme
-                # for testing frontend purposes Only
-                # log = Log(operation='+', operand='Company', user=request.user)
-                # log.add_or_remove_fields(company)
-                # log.save()
-                return Response(status=status.HTTP_200_OK)
+                log = Log(operation='+', operand='Company', user=request.user)
+                log.add_or_remove_fields(company)
+                log.save()
+                return Response(status=status.HTTP_201_CREATED)
         except serializers.ValidationError as e:
-            return Response(e.detail, status=status.HTTP_400_BAD_REQUEST)
+            return Response(e.detail, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
+        except KeyError:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
 
 
 class GetUsers(ListAPIView):
-    # fixme
-    # for testing frontend purposes Only
-    authentication_classes = []
-    permission_classes = (AllowAny,)
-    serializer_class = UserLoginSerializer
+    serializer_class = UserProfileSerializer
     queryset = User.objects.all()
 
 
 # noinspection PyMethodMayBeStatic
 class Profile(APIView):
-    # fixme
-    # for testing frontend purposes Only
-    authentication_classes = []
-    permission_classes = (AllowAny,)
-
     @transaction.atomic
     def put(self, request, pk):
         try:
             with transaction.atomic():
                 user = User.objects.get(pk=pk)
                 is_superuser = request.user.is_superuser
-                # if request.user.pk != pk and not is_superuser:
-                #     return Response(status=status.HTTP_400_BAD_REQUEST)
-                # elif is_superuser:
-                data = request.data
-                if data['password'] == '':
-                    data.pop('password')
-                address = user.address
-                address_serializer = AddressSerializer(address, data=data['address'])
-                address_serializer.is_valid(raise_exception=True)
-                address_serializer.save()
-                data.pop('address')
-                serializer = UserSerializer(user, data=data, partial=True)
-                serializer.is_valid(raise_exception=True)
-                new_user = serializer.save()
-                # fixme
-                # for testing frontend purposes Only
-                # log = Log(user=request.user, operation='*', operand='User')
-                # log.edit_fields(user, new_user)
-                # log.save()
-                return Response(serializer.data, status=status.HTTP_200_OK)
-                # else:
-                #     data = {
-                #         'password': request.data['password']
-                #     }
-                #     serializer = UserSerializer(user, data=data, partial=True)
-                #     serializer.is_valid(raise_exception=True)
-                #     serializer.save()
-                #     return Response(serializer.data, status=status.HTTP_200_OK)
+                if request.user.pk != pk and not is_superuser:
+                    return Response(status=status.HTTP_403_FORBIDDEN)
+                elif is_superuser:
+                    old_user = User.objects.get(pk=pk)
+                    data = request.data
+                    if data['password'] == '':
+                        data.pop('password')
+                    address = user.address
+                    address_serializer = AddressSerializer(address, data=data['address'])
+                    address_serializer.is_valid(raise_exception=True)
+                    address_serializer.save()
+                    data.pop('address')
+                    serializer = UserSerializer(user, data=data, partial=True)
+                    serializer.is_valid(raise_exception=True)
+                    new_user = serializer.save()
+                    log = Log(user=request.user, operation='*', operand='User')
+                    log.edit_fields(old_user, new_user)
+                    log.save()
+                    return Response(status=status.HTTP_204_NO_CONTENT)
+                else:
+                    data = {
+                        'password': request.data['password']
+                    }
+                    serializer = UserSerializer(user, data=data, partial=True)
+                    serializer.is_valid(raise_exception=True)
+                    serializer.save()
+                    return Response(status=status.HTTP_204_NO_CONTENT)
         except serializers.ValidationError as e:
-            print(e)
-            return Response(e.detail, status=status.HTTP_400_BAD_REQUEST)
+            return Response(e.detail, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
         except User.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
+        except KeyError:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
 
     def get(self, request, pk):
         try:
-            # if request.user.id != pk and not request.user.is_superuser:
-            #     return Response(status=status.HTTP_401_UNAUTHORIZED)
+            if request.user.id != pk and not request.user.is_superuser:
+                return Response(status=status.HTTP_403_FORBIDDEN)
             user = User.objects.get(id=pk)
-            serializer = UserLoginSerializer(user)
+            serializer = UserProfileSerializer(user)
             return Response(serializer.data)
         except User.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
@@ -352,32 +317,26 @@ class Profile(APIView):
             if request.user.pk == pk or not is_superuser:
                 return Response(status=status.HTTP_403_FORBIDDEN)
             else:
-                # fixme
-                # for testing frontend purposes Only
-                # log = Log(user=request.user, operation='-', operand='User')
-                # log.add_or_remove_fields(user)
-                # log.save()
+                log = Log(user=request.user, operation='-', operand='User')
+                log.add_or_remove_fields(user)
+                log.save()
                 user.delete()
-                return Response(status=status.HTTP_200_OK)
+                return Response(status=status.HTTP_204_NO_CONTENT)
         except User.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
 
-# -----------------------------------------------------------------------
 # noinspection PyMethodMayBeStatic
 class GetLog(APIView):
-    # fixme
-    # for testing frontend purposes Only
-    authentication_classes = []
-    permission_classes = (AllowAny,)
-
     def get(self, request, pk):
         try:
+            if request.user.id != pk and not request.user.is_superuser:
+                return Response(status=status.HTTP_403_FORBIDDEN)
             user = User.objects.get(pk=pk)
             logs = Log.objects.filter(user=user)
             serializer = LogSerializer(logs, many=True)
             if not serializer.data:
-                return Response({'error: ': 'No logs found!'}, status=status.HTTP_200_OK)
+                return Response(status=status.HTTP_204_NO_CONTENT)
             return Response(serializer.data, status=status.HTTP_200_OK)
         except User.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
@@ -385,18 +344,11 @@ class GetLog(APIView):
 
 # noinspection PyMethodMayBeStatic
 class UserImage(APIView):
-    # fixme
-    # for testing frontend purposes Only
-    authentication_classes = []
-    permission_classes = (AllowAny,)
-
     def get(self, request, pk):
         try:
             user = User.objects.get(pk=pk)
-            # fixme
-            # for testing frontend purposes Only
-            # if request.user.pk != request.pk and not request.user.is_superuser:
-            #     return Response(status=status.HTTP_401_UNAUTHORIZED)
+            if request.user.pk != pk and not request.user.is_superuser:
+                return Response(status=status.HTTP_403_FORBIDDEN)
             return FileResponse(user.profile_pic)
         except User.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
@@ -404,18 +356,26 @@ class UserImage(APIView):
     def put(self, request, pk):
         try:
             user = User.objects.get(pk=pk)
+            if request.user.pk != pk and not request.user.is_superuser:
+                return Response(status=status.HTTP_403_FORBIDDEN)
             if user.profile_pic.name != 'default.png':
                 user.profile_pic.delete()
             user.profile_pic = request.data["profile_pic"]
             user.save()
-            return Response(status=status.HTTP_200_OK)
+            return Response(status=status.HTTP_204_NO_CONTENT)
         except User.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
     def delete(self, request, pk):
         try:
             user = User.objects.get(pk=pk)
-            user.profile_pic.delete()
-            return Response(status=status.HTTP_200_OK)
+            if request.user.pk != pk and not request.user.is_superuser:
+                return Response(status=status.HTTP_403_FORBIDDEN)
+            if user.profile_pic.name != 'default.png':
+                user.profile_pic.delete()
+                user.profile_pic = None
+                user.save()
+                # Todo Test
+            return Response(status=status.HTTP_204_NO_CONTENT)
         except User.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
